fix: Unable to transfer files again after network disconnection
Change to a short connection, send once and close TCP once. Modifying the issue of sending data errors

Log: Unable to transfer files again after network disconnection

+++ 3rdparty/zrpc/src/net/tcpbuffer.cpp
@@ -116,8 +116,8 @@ void TcpBuffer::recycleWrite(int index) {
 }
 
 std::string TcpBuffer::getBufferString() {
-    std::string re(readAble(), '0');
-    memcpy(&re[0], &m_buffer[m_read_index], readAble());
+    std::string re(static_cast<size_t>(readAble()), '0');
+    memcpy(&re[0], &m_buffer[static_cast<size_t>(m_read_index)], static_cast<size_t>(readAble()));
     return re;
 }
